pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/foundry_bridge"]

[project]
name = "foundry-bridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "websockets>=16.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
]

[project.optional-dependencies]
transcriber = [
    "deepgram-sdk>=3.0",
]

[project.scripts]
foundry-bridge = "foundry_bridge.server:main"
foundry-bridge-transcriber = "foundry_bridge.subscribers.transcriber:main"
foundry-bridge-example = "foundry_bridge.subscribers.example:main"
